cases components abstraction
