--- src/pages/Home/components/Cases/Cases.scss
@@ -1,38 +1,0 @@
-#cases {
-    min-height: 519px;
-    background-color: $light_blue;
-    > .wrapper {
-        padding: 20px;
-        > .title {
-            margin-top: 58px;
-            margin-bottom: 77px;
-            text-align: center;
-            h2 {
-                @include below(1024px){
-                    font-family: Open Sans;
-                    font-style: normal;
-                    font-weight: 600;
-                    font-size: 18px;
-                }
-                font-size: 32px;
-            }
-        }
-
-        > .cases-data {
-            display: flex;
-            justify-content: space-evenly;
-            max-width: 1240px;
-            margin: 0 auto;
-            @include below(1024px){
-                flex-direction: column;
-            }
-            margin-bottom: 75px;
-        }
-
-        > .button-wrapper {
-            width: 100%;
-            text-align: center;
-            margin-bottom: 63px;
-        }
-    }
-}
