+++ src/pages/Home/components/CasesContainer/CasesContainer.scss
@@ -1,0 +1,27 @@
+#cases {
+  min-height: 519px;
+  background-color: $light_blue;
+  > .wrapper {
+      padding: 20px;
+      > .title {
+          margin-top: 58px;
+          margin-bottom: 77px;
+          text-align: center;
+          h2 {
+              @include below(1024px){
+                  font-family: Open Sans;
+                  font-style: normal;
+                  font-weight: 600;
+                  font-size: 18px;
+              }
+              font-size: 32px;
+          }
+      }
+
+      > .button-wrapper {
+          width: 100%;
+          text-align: center;
+          margin-bottom: 63px;
+      }
+  }
+}
