+++ src/pages/Home/index.jsx
@@ -1,16 +1,16 @@
 import Layout from "layouts"
 
 import Apply from "./components/Apply"
+import CasesContainer from './components/CasesContainer';
 import Footer from "./components/Footer"
 import Hero from "./components/Hero"
 import TalkWithUs from "./components/TalkWithUs"
-import Cases from "./components/Cases"
 
 function Home() {
   return (
     <Layout>
       <Hero />
-      <Cases />
+      <CasesContainer />
       <Apply />
 
       <TalkWithUs />
